Separated commands into their own files. Generalised provisioning message.
