+++ src/gluon/commands/packages/ConfigurePackage.ts
@@ -23,7 +23,6 @@ import {
     ApplicationType,
     menuForApplications,
 } from "../../util/packages/Applications";
-import {PackageDefinition} from "../../util/packages/PackageDefinition";
 import {getProjectDevOpsId, getProjectId} from "../../util/project/Project";
 import {
     menuForProjects,
@@ -43,112 +42,6 @@ import {
 import {TenantService} from "../../util/shared/TenantService";
 import {menuForTeams, TeamService} from "../../util/team/TeamService";
 import {KickOffJenkinsBuild} from "../jenkins/JenkinsBuild";
-
-@CommandHandler("Configure an existing application/library using a predefined template", QMConfig.subatomic.commandPrefix + " configure package")
-export class ConfigureBasicPackage extends RecursiveParameterRequestCommand {
-    @MappedParameter(MappedParameters.SlackUserName)
-    public screenName: string;
-
-    @MappedParameter(MappedParameters.SlackChannelName)
-    public teamChannel: string;
-
-    @Parameter({
-        description: "application name",
-        required: false,
-        displayable: false,
-    })
-    public applicationName: string;
-
-    @Parameter({
-        description: "project name",
-        required: false,
-        displayable: false,
-    })
-    public projectName: string;
-
-    @Parameter({
-        description: "team name",
-        required: false,
-        displayable: false,
-    })
-    public teamName: string;
-
-    @RecursiveParameter({
-        description: "package definition file",
-    })
-    public packageDefinition: string;
-
-    private readonly PACKAGE_DEFINITION_EXTENSION = ".json";
-    private readonly PACKAGE_DEFINITION_FOLDER = "resources/package-definitions/";
-
-    protected async runCommand(ctx: HandlerContext): Promise<HandlerResult> {
-        try {
-            return await this.callPackageConfiguration(ctx);
-        } catch (error) {
-            return await handleQMError(new ResponderMessageClient(ctx), error);
-        }
-    }
-
-    protected async setNextParameter(ctx: HandlerContext): Promise<HandlerResult> {
-        if (_.isEmpty(this.packageDefinition)) {
-            return await this.requestPackageDefinitionFile(ctx);
-        }
-    }
-
-    private async requestPackageDefinitionFile(ctx: HandlerContext): Promise<HandlerResult> {
-        const fs = require("fs");
-        const packageDefinitionOptions: string [] = [];
-        logger.info(`Searching folder: ${this.PACKAGE_DEFINITION_FOLDER}`);
-        fs.readdirSync(this.PACKAGE_DEFINITION_FOLDER).forEach(file => {
-            logger.info(`Found file: ${file}`);
-            if (file.endsWith(this.PACKAGE_DEFINITION_EXTENSION)) {
-                packageDefinitionOptions.push(this.getNameFromDefinitionPath(file));
-            }
-        });
-        return await createMenu(ctx, packageDefinitionOptions.map(packageDefinition => {
-                return {
-                    value: packageDefinition,
-                    text: packageDefinition,
-                };
-            }),
-            this,
-            "Please select a package definition to use for your project",
-            "Select a package definition",
-            "packageDefinition");
-    }
-
-    private async callPackageConfiguration(ctx: HandlerContext): Promise<HandlerResult> {
-        const configTemplate: QMTemplate = new QMTemplate(this.getPathFromDefinitionName(this.packageDefinition));
-        const definition: PackageDefinition = JSON.parse(configTemplate.build(QMConfig.publicConfig()));
-
-        const configurePackage = new ConfigurePackage();
-        configurePackage.screenName = this.screenName;
-        configurePackage.teamChannel = this.teamChannel;
-        configurePackage.openshiftTemplate = definition.openshiftTemplate || "";
-        configurePackage.jenkinsfileName = definition.jenkinsfile;
-        configurePackage.baseS2IImage = definition.buildConfig.imageStream;
-        if (definition.buildConfig.envVariables != null) {
-            configurePackage.buildEnvironmentVariables = definition.buildConfig.envVariables;
-        }
-        configurePackage.applicationName = this.applicationName;
-        configurePackage.teamName = this.teamName;
-        configurePackage.projectName = this.projectName;
-
-        return await configurePackage.handle(ctx);
-    }
-
-    private getNameFromDefinitionPath(definitionPath: string): string {
-        const definitionSlashSplit = definitionPath.split("/");
-        let name = definitionSlashSplit[definitionSlashSplit.length - 1];
-        // Remove file extension
-        name = name.substring(0, definitionPath.length - this.PACKAGE_DEFINITION_EXTENSION.length);
-        return name;
-    }
-
-    private getPathFromDefinitionName(definitionName: string): string {
-        return this.PACKAGE_DEFINITION_FOLDER + definitionName + this.PACKAGE_DEFINITION_EXTENSION;
-    }
-}
 
 @CommandHandler("Configure an existing application/library", QMConfig.subatomic.commandPrefix + " configure custom package")
 export class ConfigurePackage extends RecursiveParameterRequestCommand {
@@ -493,47 +386,26 @@ export class ConfigurePackage extends RecursiveParameterRequestCommand {
             logger.info(`Found tenant: ${tenant}`);
             await this.createApplicationOpenshiftResources(tenant.name, project.name, packageName);
 
-            return await this.sendApplicationProvisionedMessage(ctx, packageName, projectName, associatedTeams);
+            return await this.sendPackageProvisionedMessage(ctx, packageName, projectName, associatedTeams, ApplicationType.DEPLOYABLE);
         } else {
-            return await this.sendLibraryProvisionedMessage(ctx, packageName, projectName, associatedTeams);
+            return await this.sendPackageProvisionedMessage(ctx, packageName, projectName, associatedTeams, ApplicationType.LIBRARY);
         }
     }
 
-    private async sendApplicationProvisionedMessage(ctx: HandlerContext, applicationName: string, projectName: string, associatedTeams: any[]) {
-        return await ctx.messageClient.addressChannels({
-            text: `Your application *${applicationName}*, in project *${projectName}*, has been provisioned successfully ` +
-            "and is ready to build and deploy to your project environments",
-            attachments: [{
-                fallback: `Your application has been provisioned successfully`,
-                footer: `For more information, please read the ${this.docs() + "#jenkins-build"}`,
-                text: `
-You can kick off the build pipeline for your application by clicking the button below or pushing changes to your application's repository`,
-                mrkdwn_in: ["text"],
-                actions: [
-                    buttonForCommand(
-                        {
-                            text: "Start build",
-                            style: "primary",
-                        },
-                        new KickOffJenkinsBuild(),
-                        {
-                            projectName,
-                            applicationName,
-                        }),
-                ],
-            }],
-        }, associatedTeams.map(team =>
-            team.slack.teamChannel));
-    }
+    private async sendPackageProvisionedMessage(ctx: HandlerContext, applicationName: string, projectName: string, associatedTeams: any[], applicationType: ApplicationType) {
+        let packageTypeString = "application";
+        if (applicationType === ApplicationType.LIBRARY) {
+            packageTypeString = "library";
+        }
 
-    private async sendLibraryProvisionedMessage(ctx: HandlerContext, applicationName: string, projectName: string, associatedTeams: any[]) {
         return await ctx.messageClient.addressChannels({
-            text: "Your library has been provisioned successfully and is ready to build",
+            text: `Your ${packageTypeString} *${applicationName}*, in project *${projectName}*, has been provisioned successfully ` +
+            "and is ready to build/deploy",
             attachments: [{
-                fallback: `Your library has been provisioned successfully`,
+                fallback: `Your ${packageTypeString} has been provisioned successfully`,
                 footer: `For more information, please read the ${this.docs() + "#jenkins-build"}`,
                 text: `
-You can kick off the build pipeline for your library by clicking the button below or pushing changes to your library's repository`,
+You can kick off the build pipeline for your ${packageTypeString} by clicking the button below or pushing changes to your ${packageTypeString}'s repository`,
                 mrkdwn_in: ["text"],
                 actions: [
                     buttonForCommand(
